Refactor serialization.
